Run the collector as a command
